Remove storage hierarchy structure to use single entity storage (#128)
