--- aes-core/src/main/java/com/ijioio/aes/core/entity/storage/logical/LogicalTransaction.java
@@ -1,12 +1,0 @@
-package com.ijioio.aes.core.entity.storage.logical;
-
-import com.ijioio.aes.core.entity.storage.StorageException;
-
-public interface LogicalTransaction {
-
-	public void begin() throws StorageException;
-
-	public void commit() throws StorageException;
-
-	public void rollback() throws StorageException;
-}
